Backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.9.0
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
platformdirs==2.6.2
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
virtualenv==20.17.1
